perf: 修改作业模板的全局变量默认值，希望同步到执行方案 #3954
1. 重构TaskPlanServiceImpl，职责抽象
