+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/plan/TaskPlanService.java
@@ -211,24 +211,6 @@ public interface TaskPlanService {
     Boolean checkPlanName(Long appId, Long templateId, Long planId, String name);
 
     /**
-     * 同步执行方案
-     *
-     * @param appId           业务 ID
-     * @param templateId      模版 ID
-     * @param planId          执行方案 ID
-     * @param templateVersion 模版版本
-     * @return 是否符合条件
-     */
-    Boolean sync(Long appId, Long templateId, Long planId, String templateVersion);
-
-    /**
-     * 同步作业模版与执行方案
-     *
-     * @param taskPlan 执行方案信息
-     */
-    void syncPlan(TaskPlanInfoDTO taskPlan);
-
-    /**
      * 新增执行方案（仅数据迁移用）
      *
      * @param user           用户账号
@@ -307,7 +289,5 @@ public interface TaskPlanService {
     Integer countTaskPlans(Long appId);
 
     Set<String> listLocalFiles();
-
-    boolean batchUpdatePlanVariable(List<TaskPlanInfoDTO> planInfoList);
 
 }
